Build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@65734 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/DrvAudioVRDE.cpp
@@ -507,7 +507,7 @@ static DECLCALLBACK(uint32_t) drvAudioVRDEStreamGetReadable(PPDMIHOSTAUDIO pInte
     {
         /* Return samples instead of bytes here
          * (since we specified PDMAUDIOSTREAMLAYOUT_RAW as the audio data layout). */
-        return PDMAUDIOSTREAMCFG_B2S(pStreamVRDE->pCfg, RTCircBufUsed(pStreamVRDE->In.pCircBuf));
+        return (uint32_t)PDMAUDIOSTREAMCFG_B2S(pStreamVRDE->pCfg, RTCircBufUsed(pStreamVRDE->In.pCircBuf));
     }
 
     return 0;
